docs updated
