+++ docs/BaseRequestHandler.html
@@ -336,7 +336,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line327">line 327</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line258">line 258</a>
     </li></ul></dd>
     
 
@@ -1027,7 +1027,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line296">line 296</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line241">line 241</a>
     </li></ul></dd>
     
 
@@ -1205,7 +1205,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line147">line 147</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line185">line 185</a>
     </li></ul></dd>
     
 
@@ -1252,6 +1252,94 @@
 </dl>
 
     
+
+
+
+
+
+        
+            
+
+    
+
+    
+    <h4 class="name" id="getHeaders"><span class="type-signature"></span>getHeaders<span class="signature">()</span><span class="type-signature"></span></h4>
+    
+
+    
+
+
+
+<div class="description">
+    <p>Gets authentication common header for HTTP requests</p>
+</div>
+
+
+
+
+
+
+
+
+
+
+
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line121">line 121</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 
 
@@ -1418,7 +1506,7 @@ circumstances.</p></td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line197">line 197</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line204">line 204</a>
     </li></ul></dd>
     
 
@@ -1617,7 +1705,7 @@ if not required.</p></td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line248">line 248</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line224">line 224</a>
     </li></ul></dd>
     
 
@@ -1776,7 +1864,7 @@ if not required.</p></td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line125">line 125</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line163">line 163</a>
     </li></ul></dd>
     
 
@@ -1906,7 +1994,7 @@ if not required.</p></td>
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Wed Nov 08 2023 20:24:45 GMT-0400 (Venezuela Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Mon Nov 13 2023 11:28:33 GMT-0400 (Venezuela Time)
 </footer>
 
 <script> prettyPrint(); </script>
