+++ docs/SQLPassThruQuery.html
@@ -1254,7 +1254,7 @@ for most methods.</p></div>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line296">line 296</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line241">line 241</a>
     </li></ul></dd>
     
 
@@ -1437,7 +1437,7 @@ for most methods.</p></div>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line147">line 147</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line185">line 185</a>
     </li></ul></dd>
     
 
@@ -1484,6 +1484,99 @@ for most methods.</p></div>
 </dl>
 
     
+
+
+
+
+
+        
+            
+
+    
+
+    
+    <h4 class="name" id="getHeaders"><span class="type-signature"></span>getHeaders<span class="signature">()</span><span class="type-signature"></span></h4>
+    
+
+    
+
+
+
+<div class="description">
+    <p>Gets authentication common header for HTTP requests</p>
+</div>
+
+
+
+
+
+
+
+
+
+
+
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-overrides">Overrides:</dt>
+    <dd class="tag-overrides"><ul class="dummy"><li>
+        <a href="BaseRequestHandler.html#getHeaders">BaseRequestHandler#getHeaders</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line121">line 121</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 
 
@@ -1655,7 +1748,7 @@ circumstances.</p></td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line197">line 197</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line204">line 204</a>
     </li></ul></dd>
     
 
@@ -1859,7 +1952,7 @@ if not required.</p></td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line248">line 248</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line224">line 224</a>
     </li></ul></dd>
     
 
@@ -2023,7 +2116,7 @@ if not required.</p></td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line125">line 125</a>
+        <a href="baserequesthandler.js.html">baserequesthandler.js</a>, <a href="baserequesthandler.js.html#line163">line 163</a>
     </li></ul></dd>
     
 
@@ -2153,7 +2246,7 @@ if not required.</p></td>
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Wed Nov 08 2023 20:24:45 GMT-0400 (Venezuela Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Mon Nov 13 2023 11:28:33 GMT-0400 (Venezuela Time)
 </footer>
 
 <script> prettyPrint(); </script>
