+++ docs/baserequesthandler.js.html
@@ -143,6 +143,44 @@ class BaseRequestHandler {
         return this;
     }
 
+    /**
+    * Gets authentication common header for HTTP requests
+    */
+    getHeaders(){
+        let headers = {};
+        if (this.sdbClient.apiKey) {
+            headers = { 
+                apiKey: this.sdbClient.apiKey, 
+                Accept: this.acceptHeader,
+                'Content-Type': this.contentTypeHeader,
+                ...this.extraHeaders
+            };
+        } else if (this.sdbClient.basic) {
+            headers = { 
+                Authorization: "Basic " + this.sdbClient.basic, 
+                Accept: this.acceptHeader,
+                'Content-Type': this.contentTypeHeader,
+                ...this.extraHeaders
+            };
+        } else if (this.sdbClient.ssoCredentials) {
+            const token = btoa(this.sdbClient.ssoCredentials.id_token)
+            headers = { 
+                Authorization: "Bearer " + token,
+                "X-Identity-Provider-Id": this.sdbClient.sso.idpId, 
+                Accept: this.acceptHeader,
+                'Content-Type': this.contentTypeHeader,
+                ...this.extraHeaders
+            };
+        } else {
+            headers = { 
+                Accept: this.acceptHeader,
+                'Content-Type': this.contentTypeHeader,
+                ...this.extraHeaders
+            };
+        }
+        return headers;
+    }
+
    /**
    * Sets arbitrary custom header value for HTTP requests
    * @param {object} headersObj - an object containing key/value pairs of header properties
@@ -174,38 +212,7 @@ class BaseRequestHandler {
    */    
     async get(path) {
         const url = this._buildEndpointString(path);
-
-        let headers = {};
-        if (this.sdbClient.apiKey) {
-            headers = { 
-                apiKey: this.sdbClient.apiKey, 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        } else if (this.sdbClient.basic) {
-            headers = { 
-                Authorization: "Basic " + this.sdbClient.basic, 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        } else if (this.sdbClient.ssoCredentials) {
-            const token = btoa(this.sdbClient.ssoCredentials.id_token)
-            headers = { 
-                Authorization: "Bearer " + token,
-                "X-Identity-Provider-Id": this.sdbClient.idpId, 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        } else {
-            headers = { 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        }
+        let headers = this.getHeaders();
 
         return fetchWrapper('GET', url, undefined, headers);
     }
@@ -228,38 +235,7 @@ class BaseRequestHandler {
         }        
 
         const url = this._buildEndpointString(path);
-
-        let headers = {};
-        if (this.sdbClient.apiKey) {
-            headers = { 
-                apiKey: this.sdbClient.apiKey, 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        } else if (this.sdbClient.basic) {
-            headers = { 
-                Authorization: "Basic " + this.sdbClient.basic, 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        } else if (this.sdbClient.ssoCredentials) {
-            const token = btoa(this.sdbClient.ssoCredentials.id_token)
-            headers = { 
-                Authorization: "Bearer " + token,
-                "X-Identity-Provider-Id": this.sdbClient.idpId,
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        } else {
-            headers = { 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        }
+        let headers = this.getHeaders();
 
         return fetchWrapper('POST', url, data, headers);
     }
@@ -279,38 +255,7 @@ class BaseRequestHandler {
         }    
 
         const url = this._buildEndpointString(path);
-
-        let headers = {};
-        if (this.sdbClient.apiKey) {
-            headers = { 
-                apiKey: this.sdbClient.apiKey, 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        } else if (this.sdbClient.basic) {
-            headers = { 
-                Authorization: "Basic " + this.sdbClient.basic, 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        } else if (this.sdbClient.ssoCredentials) {
-            const token = btoa(this.sdbClient.ssoCredentials.id_token)
-            headers = { 
-                Authorization: "Bearer " + token,
-                "X-Identity-Provider-Id": this.sdbClient.idpId,
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        } else {
-            headers = { 
-                Accept: this.acceptHeader,
-                'Content-Type': this.contentTypeHeader,
-                ...this.extraHeaders
-            };
-        }
+        let headers = this.getHeaders();
 
         return fetchWrapper('PUT', url, data, headers);
     }    
@@ -323,21 +268,7 @@ class BaseRequestHandler {
    */       
     async delete(path) {
         const url = this._buildEndpointString(path);
-        
-        let headers = {};
-        if (this.sdbClient.apiKey) {
-            headers = { 
-                apiKey: this.sdbClient.apiKey, 
-                Accept: this.acceptHeader,
-                ...this.extraHeaders
-            };
-        }
-        else {
-            headers = { 
-                Accept: this.acceptHeader,
-                ...this.extraHeaders
-            };
-        }
+        let headers = this.getHeaders();
 
         return fetchWrapper('DELETE', url, undefined, headers);
     }
@@ -399,7 +330,7 @@ export { BaseRequestHandler }</code></pre>
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Wed Nov 08 2023 20:24:45 GMT-0400 (Venezuela Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Mon Nov 13 2023 11:28:33 GMT-0400 (Venezuela Time)
 </footer>
 
 <script> prettyPrint(); </script>
