+++ docs/datadiscoveryfilter.js.html
@@ -325,7 +325,7 @@ export { SDB_DDF_INVALID_WILDCARD, SDB_DDF_INVALID_SEPARATOR,
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Wed Nov 08 2023 20:24:45 GMT-0400 (Venezuela Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.2</a> on Mon Nov 13 2023 11:28:33 GMT-0400 (Venezuela Time)
 </footer>
 
 <script> prettyPrint(); </script>
