+++ src/slashdbclient.js
@@ -1,9 +1,8 @@
 import { DataDiscoveryDatabase } from './datadiscovery.js'
 import { SQLPassThruQuery } from './sqlpassthru.js'
 import { BaseRequestHandler } from './baserequesthandler.js';
-import { isSSOredirect } from './sso.js';
 import { PKCE } from './pkce.js';
-import { generateCodeChallenge, generateRandomString, getUrlParms, popupCenter } from "./utils.js";
+import { generateCodeChallenge, generateRandomString, getUrlParms, isSSOredirect, popupCenter } from "./utils.js";
 
 const SDB_SDBC_INVALID_HOSTNAME = 'Invalid hostname parameter, must be string';
 const SDB_SDBC_INVALID_USERNAME = 'Invalid username parameter, must be string';
