+++ src/utils.js
@@ -61,4 +61,34 @@ const popupCenter = (url, title, w, h) => {
     return newWindow;
 }
 
-export { generateRandomString, generateCodeChallenge, getUrlParms, popupCenter }
+function isObjectEmpty(objectName) {
+    return Object.keys(objectName).length === 0
+}
+
+function isSSOredirect(ssoParams) {
+
+    if (isObjectEmpty(ssoParams)){
+        return false;
+    }
+
+    const state = ssoParams.state;
+    const sessionState = ssoParams.session_state;
+    const code = ssoParams.code;
+
+    if (!state || typeof(state) !== 'string') {
+        return false
+    }
+
+    if (!sessionState || typeof(sessionState) !== 'string') {
+        return false
+    }
+
+    if (!code || typeof(code) !== 'string') {
+        return false
+    }
+
+    return true;
+
+}
+
+export { generateRandomString, generateCodeChallenge, getUrlParms, popupCenter, isSSOredirect }
